evocode/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evocode"
version = "0.1.0"
description = "Evolutionary Code Generation Library for Python"
readme = "README.md"
authors = [
    {name = "EvoCode Team", email = "info@evocode.example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "matplotlib>=3.3.0",
    "networkx>=2.5",
    "numpy>=1.19.0",
    "python-Levenshtein>=0.12.0",
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/evocode-team/evocode"
"Bug Tracker" = "https://github.com/evocode-team/evocode/issues"
"Documentation" = "https://evocode.readthedocs.io"

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=20.8b1",
    "flake8>=3.8.0",
    "mypy>=0.790",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88 